# PeopleCounter_Manager/docs/swagger.yaml
basePath: /
definitions:
...
  PeopleCounter_Manager_internal_models.LoginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  PeopleCounter_Manager_internal_models.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  PeopleCounter_Manager_internal_models.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  PeopleCounter_Manager_internal_models.User:
    properties:
      description:
        type: string
      email:
        type: string
      id:
        type: integer
      login:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
...
host: localhost:9000
info:
  contact:
    email: support@example.com
    name: API Support
  description: REST API для управления счетчиками посетителей и сбора статистики
  termsOfService: http://swagger.io/terms/
  title: People Counter Manager API
  version: "1.0"
paths:
...
  /api/login:
    post:
      consumes:
      - application/json
      description: Аутентифицирует пользователя по логину и паролю и возвращает пару
        JWT токенов (access и refresh).
      parameters:
      - description: Учетные данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/PeopleCounter_Manager_internal_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная авторизация
          schema:
            $ref: '#/definitions/PeopleCounter_Manager_internal_models.TokenResponse'
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "401":
          description: Неверный логин или пароль
          schema:
            type: string
      summary: Вход в систему (Авторизация)
      tags:
      - Auth
  /api/refresh:
    post:
      consumes:
      - application/json
      description: Принимает действующий refresh токен и возвращает новую пару токенов
        (access и refresh).
      parameters:
      - description: Refresh токен
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/PeopleCounter_Manager_internal_models.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Новая пара токенов
          schema:
            $ref: '#/definitions/PeopleCounter_Manager_internal_models.TokenResponse'
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "401":
          description: Недействительный или истекший refresh токен
          schema:
            type: string
      summary: Обновление токена доступа
      tags:
      - Auth
...
  /api/user:
    get:
      description: Возвращает данные текущего авторизованного пользователя на основе
        его JWT токена.
      produces:
      - application/json
      responses:
        "200":
          description: Профиль пользователя
          schema:
            $ref: '#/definitions/PeopleCounter_Manager_internal_models.User'
        "401":
          description: Не авторизован (отсутствует или неверный токен)
          schema:
            type: string
        "404":
          description: Пользователь не найден в БД
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить профиль пользователя
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"